requirements.txt
psycopg2-binary==2.9.9
fastapi==0.115.6
uvicorn==0.34.0
openai>=1.159.8